// src/types/customDataTypes.ts
export interface FacilityTypes {
  title: string;
  description: string;
  imageUrl: string;
}
// ...
export interface CustomCardGridDataTypes {
  icon: string;
  title: string;
  pdf: string;
  backgroundImage: string;
}

interface contentItem {
  M_name: string;
  M_mobNo: string;
  C_name: string;
  C_mobNo: string;
}

export interface transportationDataTypes {
  content: string;
  src: string;
  alt: string;
  contentItem: contentItem[];
}

export interface TableRow {
  slNo: number;
  nameOfBody: string;
  membershipMode: string;
  period: string;
  startFrom: string;
  membershipNo: string;
}

export interface otherFacilitiesCardLayoutDataTypes {
  title: string;
  content: string;
}